base/fastapi/pyproject.toml
[project]
name = "fastapi-base"
version = "0.1.0"
description = "FastAPI starter kit — DDD + JWT + Read/Write DB + S3"
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "boto3>=1.40.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "gunicorn>=23.0.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "pyjwt>=2.10.0",
    "python-ulid>=3.1.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.25",
    "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "alembic>=1.17.0",
    "black>=25.0.0",
    "isort>=6.0.0",
    "mypy>=1.18.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.0",
    "httpx>=0.27.0",
]
